Split `cluster` package into 3 files
